backend/notes_crud_service/requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
mangum
python-jose
google-generativeai
langchain-google-genai